Clean up stage bindings, wrap in proper types.
This changes the stage bindings to use a proper stage
handle, rather than passing a path to keep reopening stages.
The stage is kept alive by being statically constructed in
_openStage(). We'll need to put something in there for
destructing the stage once we provide an option to reload
or open a new stage in the UI.

This change also encapsulates the information inside the
Swift type, so users get simple property value types to
interact with.

Towards #10